feat: 添加 download 命令的解释模式
