+++ src/kmdr/module/downloader/DirectDownloader.py
@@ -1,23 +1,25 @@
 from functools import partial
 from typing import Callable, Optional
 
-from kmdr.core import DOWNLOADER, BookInfo, Downloader, VolInfo
+from kmdr.core import DOWNLOADER, BookInfo, VolInfo
 from kmdr.core.constants import API_ROUTE
 from kmdr.core.structure import Credential
 
+from .base import BaseDownloader
 from .download_utils import (
     download_file,
     download_file_multipart,
+    format_filename,
     readable_safe_filename,
 )
 
 
 @DOWNLOADER.register(hasvalues={"method": 2})
-class DirectDownloader(Downloader):
+class DirectDownloader(BaseDownloader):
     def __init__(
-        self, dest=".", format="epub", callback=None, retry=3, num_workers=8, proxy=None, vip=False, disable_multi_part=False, *args, **kwargs
+        self, dest=".", format="epub", callback=None, retry=3, num_workers=8, vip=False, disable_multi_part=False, *args, **kwargs
     ):
-        super().__init__(dest, format, callback, retry, num_workers, proxy, *args, **kwargs)
+        super().__init__(dest, format, callback, retry, num_workers, *args, **kwargs)
         self._use_vip = vip
         self._disable_multi_part = disable_multi_part
 
@@ -39,7 +41,7 @@ class DirectDownloader(Downloader):
                 self._progress,
                 partial(self.construct_download_url, cred, book, volume),
                 download_path,
-                readable_safe_filename(f"[Kmoe][{book.name}][{volume.name}].{self._format.name.lower()}"),
+                format_filename(book.name, volume.name, self._format.name.lower()),
                 self._retry,
                 cookies=cred.cookies,
                 callback=lambda: self._callback(book, volume) if self._callback else None,
@@ -54,7 +56,7 @@ class DirectDownloader(Downloader):
             self._progress,
             partial(self.construct_download_url, cred, book, volume),
             download_path,
-            readable_safe_filename(f"[Kmoe][{book.name}][{volume.name}].{self._format.name.lower()}"),
+            format_filename(book.name, volume.name, self._format.name.lower()),
             self._retry,
             cookies=cred.cookies,
             callback=lambda: self._callback(book, volume) if self._callback else None,
